comment: add send & child

+++ app/src/main/java/com/duzhaokun123/bilibilihd2/ui/comment/CommentActivity.kt
@@ -25,7 +25,7 @@ class CommentActivity: BaseActivity<ActivityCommentBinding>(R.layout.activity_co
         super.initEvents()
         baseBinding.btnReload.setOnClickListener {
             rootCommentFragment.setOid(baseBinding.etOid.text.toString().toLong())
-            rootCommentFragment.setType(baseBinding.etType.text.toString().toLong())
+            rootCommentFragment.setType(baseBinding.etType.text.toString().toInt())
             rootCommentFragment.setMode(mode)
             rootCommentFragment.reload()
         }
